+++ struct/bst.h
@@ -14,7 +14,7 @@
  
  /* Binary Search Tree node structure */
  typedef struct BSTNode {
-     uint64_t key;           /* The key for the BST entry */
+     uint32_t key;           /* The key for the BST entry */
      void* value;            /* Pointer to the stored value */
      struct BSTNode* left;   /* Pointer to left child */
      struct BSTNode* right;  /* Pointer to right child */
@@ -50,7 +50,7 @@
   * @param value Pointer to the value to associate with the key.
   * @return true if insertion was successful, false if the key already exists or on failure.
   */
- bool bst_insert(BST* tree, uint64_t key, void* value);
+ bool bst_insert(BST* tree, uint32_t key, void* value);
  
  /**
   * Retrieves a value associated with a given key from the BST.
@@ -59,7 +59,7 @@
   * @param key The key to search for.
   * @return Pointer to the stored value, or NULL if the key is not found.
   */
- void* bst_get(const BST* tree, uint64_t key);
+ void* bst_get(const BST* tree, uint32_t key);
  
  /**
   * Removes a node by key from the BST.
@@ -68,7 +68,7 @@
   * @param key The key to remove.
   * @return true if the key was found and removed, false otherwise.
   */
- bool bst_remove(BST* tree, uint64_t key);
+ bool bst_remove(BST* tree, uint32_t key);
  
  /**
   * Returns the current number of nodes stored in the BST.
@@ -85,5 +85,5 @@
   * @param key The key to look for.
   * @return true if the key exists in the BST, false otherwise.
   */
- bool bst_contains(const BST* tree, uint64_t key);
+ bool bst_contains(const BST* tree, uint32_t key);
