+++ struct/hash_table.h
@@ -14,7 +14,7 @@
 
 /* Hash table node structure */
 typedef struct HashNode {
-    uint64_t key;           /* The key for the hash table entry */
+    uint32_t key;           /* The key for the hash table entry */
     void* value;            /* Pointer to the stored value */
     struct HashNode* next;  /* Pointer to next node in case of collision */
 } HashNode;
@@ -51,7 +51,7 @@ void hash_table_destroy(HashTable* table);
  * @param value The value to store.
  * @return true if the insertion was successful, false otherwise.
  */
-bool hash_table_insert(HashTable* table, uint64_t key, void* value);
+bool hash_table_insert(HashTable* table, uint32_t key, void* value);
 
 /**
  * Retrieves a value from the hash table by key.
@@ -60,7 +60,7 @@ bool hash_table_insert(HashTable* table, uint64_t key, void* value);
  * @param key The key to look up.
  * @return A pointer to the stored value, or NULL if not found.
  */
-void* hash_table_get(HashTable* table, uint64_t key);
+void* hash_table_get(HashTable* table, uint32_t key);
 
 /**
  * Removes a key-value pair from the hash table.
@@ -69,7 +69,7 @@ void* hash_table_get(HashTable* table, uint64_t key);
  * @param key The key to remove.
  * @return true if the key was found and removed, false otherwise.
  */
-bool hash_table_remove(HashTable* table, uint64_t key);
+bool hash_table_remove(HashTable* table, uint32_t key);
 
 /**
  * Returns the current size of the hash table.
@@ -86,6 +86,6 @@ size_t hash_table_size(const HashTable* table);
  * @param key The key to look for.
  * @return true if the key exists, false otherwise.
  */
-bool hash_table_contains(const HashTable* table, uint64_t key);
+bool hash_table_contains(const HashTable* table, uint32_t key);
 
 #endif // HASHTABLE_H
